pyproject.toml
[project]
name = "eng-words"
version = "0.1.0"
description = "Personal vocabulary extraction and Anki card generation tool for English learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "spacy>=3.8.11",
    "typer>=0.9.0",  # Required by spacy CLI
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "pandas>=2.0.0",
    "matplotlib>=3.10.0",
    "pyarrow>=14.0.0",  # for parquet support
    "wordfreq>=3.1.1",
    "tqdm>=4.66.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12",
    "numpy>=1.26.0",
    "gspread>=5.12.0",
    "google-auth>=2.23.0",
    "python-dotenv>=1.0.0",
    "nltk>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "isort>=5.12.0",
    "pyright>=1.1.0",
    "ipykernel>=6.25.0",
]
wsd = [
    "sentence-transformers>=5.2.0",
]
llm = [
    "openai>=2.11.0,<3",
    "anthropic>=0.75.0,<1",
    "google-genai>=1.0.0",
    "pydantic>=2.12.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--ignore=tests/_archive"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["eng_words"]

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__", "**/.*"]
pythonVersion = "3.10"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
# Ignore known pandas typing issues
reportArgumentType = false
reportReturnType = false
reportAttributeAccessIssue = false
reportCallIssue = false
reportOperatorIssue = false
reportGeneralTypeIssues = false
reportOptionalMemberAccess = false

